requirements.txt
requests==2.31.0
pandas==2.0.3
sqlalchemy==2.0.19
mysql-connector-python==8.1.0
streamlit==1.25.0
python-dotenv==1.0.0
plotly==5.15.0
argparse
